Support accessing the response code as an integer
The `Status` enum only supports a portion of the response codes defined
by RFC 9110.  If the server returned a response code other than one of
the ones supported by the `Status` enum, the code previously threw away
the value.  This updates the code to also store the code as an integer,
so that users can handle other values.

I considered simply changing the `Unknown` variant to use
`Unknown(u16)`, but avoided it for now since this would break backwards
compatibility, and would also prevent using `status as u16` to convert
`Status` values to integers.